+++ voiceDB/Voice.cpp
@@ -152,8 +152,8 @@ const UnitWaveformContainer* Voice::getUwc() const
   }
 
   // get wav data
-  vector<short> wav_data;
-  unsigned long wav_fs = 0;
+  vector<double> wav_data;
+  long wav_fs = 0;
   {
     cout << "wav data" << endl;
     WavParser wav_parser(path_wav.string());
@@ -172,10 +172,10 @@ const UnitWaveformContainer* Voice::getUwc() const
     marker->setInputWav(wav_data, offs, ovrl, prec, blnk, wav_fs);
     if (nak::pron2vow.count(pron.substr(pron.size()-2))>0 && voice_db->isVowel(nak::pron2vow[pron.substr(pron.size()-2)]+suffix)) {
       short win_size = wav_fs / getFrq() * 2;
-      vector<short> aft_vowel_wav = voice_db->getVowel(nak::pron2vow[pron.substr(pron.size()-2)]+suffix);
+      vector<double> aft_vowel_wav = voice_db->getVowel(nak::pron2vow[pron.substr(pron.size()-2)]+suffix);
       trimVector(&aft_vowel_wav, win_size);
       if (is_vcv && voice_db->isVowel(prefix.substr(0,prefix.size()-1)+suffix)) {
-        vector<short> fore_vowel_wav = voice_db->getVowel(prefix.substr(0,prefix.size()-1)+suffix);
+        vector<double> fore_vowel_wav = voice_db->getVowel(prefix.substr(0,prefix.size()-1)+suffix);
         trimVector(&fore_vowel_wav, win_size);
         marker->mark(fore_vowel_wav, aft_vowel_wav);
       } else {
